source-code/backend/requirements.txt
google-cloud-logging==3.5.0
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.18.4
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
